add tokenizer base class
