+++ src/weathergen/datasets/tokenizer_masking.py
@@ -7,13 +7,12 @@
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
 
-import warnings
 from functools import partial
 
-import astropy_healpix as hp
 import numpy as np
 import torch
 
+from weathergen.datasets.tokenizer import Tokenizer
 from weathergen.datasets.masking import Masker
 from weathergen.datasets.tokenizer_utils import (
     arc_alpha,
@@ -24,112 +23,15 @@ from weathergen.datasets.tokenizer_utils import (
 )
 from weathergen.datasets.utils import (
     get_target_coords_local_ffast,
-    healpix_verts_rots,
-    locs_to_cell_coords_ctrs,
     r3tos2,
 )
 from weathergen.utils.logger import init_loggers
 
 
-class TokenizerMasking:
+class TokenizerMasking(Tokenizer):
     def __init__(self, healpix_level: int, masker: Masker):
-        ref = torch.tensor([1.0, 0.0, 0.0])
-
-        self.hl_source = healpix_level
-        self.hl_target = healpix_level
-
+        super().__init__(healpix_level)
         self.masker = masker
-
-        self.num_healpix_cells_source = 12 * 4**self.hl_source
-        self.num_healpix_cells_target = 12 * 4**self.hl_target
-
-        verts00, verts00_rots = healpix_verts_rots(self.hl_source, 0.0, 0.0)
-        verts10, verts10_rots = healpix_verts_rots(self.hl_source, 1.0, 0.0)
-        verts11, verts11_rots = healpix_verts_rots(self.hl_source, 1.0, 1.0)
-        verts01, verts01_rots = healpix_verts_rots(self.hl_source, 0.0, 1.0)
-        vertsmm, vertsmm_rots = healpix_verts_rots(self.hl_source, 0.5, 0.5)
-        self.hpy_verts = [
-            verts00.to(torch.float32),
-            verts10.to(torch.float32),
-            verts11.to(torch.float32),
-            verts01.to(torch.float32),
-            vertsmm.to(torch.float32),
-        ]
-        self.hpy_verts_rots_source = [
-            verts00_rots.to(torch.float32),
-            verts10_rots.to(torch.float32),
-            verts11_rots.to(torch.float32),
-            verts01_rots.to(torch.float32),
-            vertsmm_rots.to(torch.float32),
-        ]
-
-        verts00, verts00_rots = healpix_verts_rots(self.hl_target, 0.0, 0.0)
-        verts10, verts10_rots = healpix_verts_rots(self.hl_target, 1.0, 0.0)
-        verts11, verts11_rots = healpix_verts_rots(self.hl_target, 1.0, 1.0)
-        verts01, verts01_rots = healpix_verts_rots(self.hl_target, 0.0, 1.0)
-        vertsmm, vertsmm_rots = healpix_verts_rots(self.hl_target, 0.5, 0.5)
-        self.hpy_verts = [
-            verts00.to(torch.float32),
-            verts10.to(torch.float32),
-            verts11.to(torch.float32),
-            verts01.to(torch.float32),
-            vertsmm.to(torch.float32),
-        ]
-        self.hpy_verts_rots_target = [
-            verts00_rots.to(torch.float32),
-            verts10_rots.to(torch.float32),
-            verts11_rots.to(torch.float32),
-            verts01_rots.to(torch.float32),
-            vertsmm_rots.to(torch.float32),
-        ]
-
-        self.verts_local = []
-        verts = torch.stack([verts10, verts11, verts01, vertsmm])
-        temp = ref - torch.stack(locs_to_cell_coords_ctrs(verts00_rots, verts.transpose(0, 1)))
-        self.verts_local.append(temp.flatten(1, 2))
-
-        verts = torch.stack([verts00, verts11, verts01, vertsmm])
-        temp = ref - torch.stack(locs_to_cell_coords_ctrs(verts10_rots, verts.transpose(0, 1)))
-        self.verts_local.append(temp.flatten(1, 2))
-
-        verts = torch.stack([verts00, verts10, verts01, vertsmm])
-        temp = ref - torch.stack(locs_to_cell_coords_ctrs(verts11_rots, verts.transpose(0, 1)))
-        self.verts_local.append(temp.flatten(1, 2))
-
-        verts = torch.stack([verts00, verts11, verts10, vertsmm])
-        temp = ref - torch.stack(locs_to_cell_coords_ctrs(verts01_rots, verts.transpose(0, 1)))
-        self.verts_local.append(temp.flatten(1, 2))
-
-        verts = torch.stack([verts00, verts10, verts11, verts01])
-        temp = ref - torch.stack(locs_to_cell_coords_ctrs(vertsmm_rots, verts.transpose(0, 1)))
-        self.verts_local.append(temp.flatten(1, 2))
-
-        self.hpy_verts_local_target = torch.stack(self.verts_local).transpose(0, 1)
-
-        # add local coords wrt to center of neighboring cells
-        # (since the neighbors are used in the prediction)
-        num_healpix_cells = 12 * 4**self.hl_target
-        with warnings.catch_warnings(action="ignore"):
-            temp = hp.neighbours(
-                np.arange(num_healpix_cells), 2**self.hl_target, order="nested"
-            ).transpose()
-        # fix missing nbors with references to self
-        for i, row in enumerate(temp):
-            temp[i][row == -1] = i
-        self.hpy_nctrs_target = (
-            vertsmm[temp.flatten()]
-            .reshape((num_healpix_cells, 8, 3))
-            .transpose(1, 0)
-            .to(torch.float32)
-        )
-
-        self.size_time_embedding = 6
-
-    def get_size_time_embedding(self) -> int:
-        """
-        Get size of time embedding
-        """
-        return self.size_time_embedding
 
     def reset_rng(self, rng) -> None:
         """
